pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.10",
    "openai>=1.101.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
